todo/requirements.txt
djangorestframework
pika